Revert "Parquet read array (#895)"
This reverts commit cca5ef88c04872eae09e61a355b0b048f3f42355.
