+++ automation/src/test/resources/data/parquet/generate_precision_numeric_parquet.hql
@@ -1,6 +1,4 @@
-DROP TABLE IF EXISTS precision_numeric_csv;
-
-CREATE EXTERNAL TABLE precision_numeric_csv (
+CREATE EXTERNAL TABLE IF NOT EXISTS precision_numeric_csv (
 description STRING,
 a DECIMAL(5,  2),
 b DECIMAL(12, 2),
