+++ automation/src/test/resources/data/parquet/generate_undefined_precision_numeric_parquet.hql
@@ -1,6 +1,4 @@
-DROP TABLE IF EXISTS undefined_precision_numeric_csv;
-
-CREATE EXTERNAL TABLE undefined_precision_numeric_csv
+CREATE EXTERNAL TABLE IF NOT EXISTS undefined_precision_numeric_csv
 (description STRING, value DECIMAL(38,18))
 ROW FORMAT
 DELIMITED FIELDS TERMINATED BY ','
