--- automation/src/test/resources/data/parquet/parquet_types.csv
@@ -1,25 +1,0 @@
-1,row1,2019-12-01,1200,good,FALSE,2013-07-13 21:00:00,2147483647,1,-32768,7.7,s_6,USD,1.23456,0,0.12345,1
-2,row2,2019-12-02,1300,excellent,TRUE,2013-07-13 21:00:00,2147483648,2,-31500,8.7,s_7,USD,1.23456,123.45,-0.12345,2
-3,row3,2019-12-03,1400,good,FALSE,2013-07-15 21:00:00,2147483649,3,-31000,9.7,s_8,USD,-1.23456,-1.45,12345678.90123,3
-4,row4,2019-12-04,1500,excellent,TRUE,2013-07-16 21:00:00,2147483650,4,-30000,10.7,s_9,USD,123456789.1,0.25,-12345678.90123,4
-5,row5,2019-12-05,1600,good,FALSE,2013-07-17 21:00:00,2147483651,5,-20000,11.7,s_10,USD,1E-12,-.25,99999999,5
-6,row6,2019-12-06,1700,bad,TRUE,2013-07-18 21:00:00,2147483652,6,-10000,12.7,s_11,USD,1234.889,999.99,-99999999,6
-7,row7,2019-12-07,1800,good,FALSE,2013-07-19 21:00:00,2147483653,7,-1000,7.7,s_12,USD,0.0001,-999.99,-99999999.99999,7
-8,row8,2019-12-08,1900,bad,TRUE,2013-07-20 21:00:00,2147483654,8,-550,7.7,s_13,EUR,45678.00002,1,99999999.99999,8
-9,row9,2019-12-09,2000,excellent,FALSE,2013-07-21 21:00:00,2147483655,9,-320,7.7,s_14,UAH,23457.1,-1,0,9
-10,row10,2019-12-10,2100,bad,TRUE,2013-07-22 21:00:00,2147483656,0,-120,7.7,s_15,USD,45678.00002,789,1,10
-11,row11,2019-12-11,2200,good,FALSE,2013-07-23 21:00:00,2147483657,1,-40,7.7,s_16,UAH,0.123456789,-789,-1,11
-12,row12_text_null,2019-12-12,2300,\N,FALSE,2013-07-23 21:00:00,2147483658,1,-1,7.7,s_16,EUR,0.123456789,0.99,0.9,11
-13,row13_int_null,2019-12-13,2400,good,FALSE,2013-07-23 21:00:00,2147483659,1,0,7.7,s_16,USD,0.123456789,-0.99,-0.9,\N
-14,row14_double_null,2019-12-14,\N,excellent,FALSE,2013-07-23 21:00:00,2147483660,1,1,7.7,s_16,UAH,0.123456789,1.99,45,11
-15,row15_decimal_null,2019-12-15,2500,good,FALSE,2013-07-24 21:00:00,2147483661,1,100,7.7,s_17,USD,\N,\N,\N,12
-16,row16_timestamp_null,2019-12-16,2550,bad,FALSE,\N,2147483662,1,1000,7.7,s_160,USD,0.123456789,-1.99,-45,11
-17,row17_real_null,2019-12-17,2600,good,FALSE,2013-07-23 21:00:00,2147483663,1,10000,\N,s_161,EUR,0.123456789,15.99,3.14159,11
-18,row18_bigint_null,2019-12-18,2600,bad,FALSE,2013-07-23 21:00:00,\N,1,20000,7.7,s_162,USD,0.123456789,-15.99,-3.14159,11
-19,row19_bool_null,2019-12-19,2600,good,\N,2013-07-23 21:00:00,-1,1,30000,7.7,s_163,USD,0.123456789,-299.99,2.71828,11
-20,row20,2019-12-20,2600,excellent,FALSE,2013-07-23 21:00:00,-2147483643,1,31000,7.7,s_164,UAH,0.123456789,299.99,-2.71828,11
-21,row21_smallint_null,2019-12-21,2600,good,FALSE,2013-07-23 21:00:00,-2147483644,1,\N,7.7,s_165,USD,0.123456789,555.55,45.99999,11
-22,row22_date_null,\N,2600,excellent,FALSE,2013-07-23 21:00:00,-2147483645,1,32100,7.7,s_166,EUR,0.123456789,0.15,-45.99999,11
-23,row23_varchar_null,2019-12-23,2600,good,FALSE,2013-07-23 21:00:00,-2147483646,1,32200,7.7,\N,EUR,0.123456789,3.89,450.45001,11
-24,row24_char_null,2019-12-24,2600,bad,FALSE,2013-07-23 21:00:00,-2147483647,1,32500,7.7,s_168,\N,0.123456789,3.14,0.00001,11
-25,row25_binary_null,2019-12-25,2600,good,FALSE,2013-07-23 21:00:00,-2147483648,\N,32767,7.7,s_169,USD,0.123456789,8,-0.00001,11
