--- automation/tincrepo/main/pxf/features/parquet/list/runTest.py
@@ -1,11 +1,0 @@
-from mpp.models import SQLConcurrencyTestCase
-
-class ParquetList(SQLConcurrencyTestCase):
-    """
-    @db_name pxfautomation
-    @concurrency 1
-    @gpdiff True
-    """
-    sql_dir = 'sql'
-    ans_dir = 'expected'
-    out_dir = 'output'
