+++ server/pxf-hdfs/src/test/java/org/greenplum/pxf/plugins/hdfs/ParquetResolverTest.java
@@ -16,12 +16,9 @@ import org.apache.parquet.io.api.Binary;
 import org.apache.parquet.pig.convert.DecimalUtils;
 import org.apache.parquet.schema.MessageType;
 import org.apache.parquet.schema.PrimitiveType;
-import org.apache.parquet.schema.GroupType;
 import org.apache.parquet.schema.Type;
 import org.greenplum.pxf.api.OneField;
 import org.greenplum.pxf.api.OneRow;
-import org.greenplum.pxf.api.error.PxfRuntimeException;
-import org.greenplum.pxf.api.error.UnsupportedTypeException;
 import org.greenplum.pxf.api.io.DataType;
 import org.greenplum.pxf.api.model.RequestContext;
 import org.greenplum.pxf.api.utilities.ColumnDescriptor;
@@ -47,7 +44,6 @@ import static org.junit.jupiter.api.Assertions.assertArrayEquals;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
-import static org.junit.jupiter.api.Assertions.assertNull;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
@@ -525,348 +521,6 @@ public class ParquetResolverTest {
 
     }
 
-    @Test
-    public void testGetFields_List() throws IOException {
-        schema = getParquetSchemaForListTypesGeneratedByHive();
-        // schema has changed, set metadata again
-        context.setMetadata(schema);
-        context.setTupleDescription(getColumnDescriptorsFromSchema(schema));
-        resolver.setRequestContext(context);
-        resolver.afterPropertiesSet();
-
-        List<Group> groups = readParquetFile("parquet_list_types_without_null.parquet", 2, schema);
-        assertEquals(2, groups.size());
-
-        List<OneField> fields = assertRow(groups, 0, 13);
-        assertField(fields, 0, 1, DataType.INTEGER);
-        assertList(fields, 1, "{true}", DataType.BOOLARRAY);
-        assertList(fields, 2, "{50}", DataType.INT2ARRAY);
-        assertList(fields, 3, "{1}", DataType.INT4ARRAY);
-        assertList(fields, 4, "{1}", DataType.INT8ARRAY);
-        assertList(fields, 5, "{1.11}", DataType.FLOAT4ARRAY);
-        assertList(fields, 6, "{1.7E308}", DataType.FLOAT8ARRAY);
-        assertList(fields, 7, "{\"this is a test string\"}", DataType.TEXTARRAY);
-        assertList(fields, 8, "{\"\\\\xdeadbeef\"}", DataType.BYTEAARRAY);
-        assertList(fields, 9, "{hello}", DataType.BPCHARARRAY);// right trimmed spaces
-        assertList(fields, 10, "{hello}", DataType.VARCHARARRAY);
-        assertList(fields, 11, "{2022-10-07}", DataType.DATEARRAY);
-        assertList(fields, 12, "{" + BigDecimal.valueOf(1234560000000000000L, 18) + "}", DataType.NUMERICARRAY);
-
-        fields = assertRow(groups, 1, 13);
-        assertField(fields, 0, 2, DataType.INTEGER);
-        assertList(fields, 1, "{false,true,true,false}", DataType.BOOLARRAY);
-        assertList(fields, 2, "{-128,96}", DataType.INT2ARRAY);
-        assertList(fields, 3, "{2,3}", DataType.INT4ARRAY);
-        assertList(fields, 4, "{-9223372036854775808,223372036854775808}", DataType.INT8ARRAY);
-        assertList(fields, 5, "{-123456.984,123456.984}", DataType.FLOAT4ARRAY);
-        assertList(fields, 6, "{1.0,-99.9}", DataType.FLOAT8ARRAY);
-        assertList(fields, 7, "{\"this is a string with \\\"special\\\" characters\",\"this is a string without\"}", DataType.TEXTARRAY);
-        assertList(fields, 8, "{\"\\\\xdeadbeef\",\"\\\\xadbeef\"}", DataType.BYTEAARRAY);
-        assertList(fields, 9, "{\"this is exactly\",\" fifteen chars.\"}", DataType.BPCHARARRAY);
-        assertList(fields, 10, "{\"this is exactly\",\" fifteen chars.\"}", DataType.VARCHARARRAY);
-        assertList(fields, 11, "{2022-10-07,2022-10-08}", DataType.DATEARRAY);
-        assertList(fields, 12, "{" + BigDecimal.valueOf(1234560000000000000L, 18) + "," + BigDecimal.valueOf(1234560000000000000L, 18) + "}", DataType.NUMERICARRAY);
-    }
-
-    @Test
-    public void testGetFields_List_Nulls() throws IOException {
-        schema = getParquetSchemaForListTypesGeneratedByHive();
-        // schema has changed, set metadata again
-        context.setMetadata(schema);
-        context.setTupleDescription(getColumnDescriptorsFromSchema(schema));
-        resolver.setRequestContext(context);
-        resolver.afterPropertiesSet();
-
-        List<Group> groups = readParquetFile("parquet_list_types.parquet", 6, schema);
-        assertEquals(6, groups.size());
-
-        List<OneField> fields = assertRow(groups, 3, 13);
-        assertField(fields, 0, 4, DataType.INTEGER);
-        assertList(fields, 1, null, DataType.BOOLARRAY);
-        assertList(fields, 2, "{10,20}", DataType.INT2ARRAY);
-        assertList(fields, 3, "{7,NULL,8}", DataType.INT4ARRAY);
-        assertList(fields, 4, "{-9223372036854775808,0}", DataType.INT8ARRAY);
-        assertList(fields, 5, "{2.3,4.5}", DataType.FLOAT4ARRAY);
-        assertList(fields, 6, null, DataType.FLOAT8ARRAY);
-        assertList(fields, 7, "{NULL,\"\"}", DataType.TEXTARRAY);
-        assertList(fields, 8, "{NULL,\"\\\\x5c22\"}", DataType.BYTEAARRAY);
-        assertList(fields, 9, null, DataType.BPCHARARRAY);
-        assertList(fields, 10, null, DataType.VARCHARARRAY);
-        assertList(fields, 11, "{2022-10-07,2022-10-07,NULL}", DataType.DATEARRAY);
-        assertList(fields, 12, "{" + BigDecimal.valueOf(1234500000000000000L, 18) + "}", DataType.NUMERICARRAY);
-    }
-
-    @Test
-    public void testGetFields_List_With_Projection() throws IOException {
-        schema = getParquetSchemaForListTypesGeneratedByHive();
-        context.setTupleDescription(getColumnDescriptorsFromSchema(schema));
-        // set odd columns to be not projected, their values will become null
-        for (int i = 0; i < context.getTupleDescription().size(); i++) {
-            context.getTupleDescription().get(i).setProjected(i % 2 == 0);
-        }
-
-        MessageType readSchema = buildReadSchema(schema);
-        // schema has changed, set metadata again
-        context.setMetadata(readSchema);
-
-        resolver.setRequestContext(context);
-        resolver.afterPropertiesSet();
-
-        List<Group> groups = readParquetFile("parquet_list_types_without_null.parquet", 2, readSchema);
-        assertEquals(2, groups.size());
-
-        List<OneField> fields = assertRow(groups, 0, 13);
-        assertField(fields, 0, 1, DataType.INTEGER);
-        assertList(fields, 1, null, DataType.BOOLARRAY);
-        assertList(fields, 2, "{50}", DataType.INT2ARRAY);
-        assertList(fields, 3, null, DataType.INT4ARRAY);
-        assertList(fields, 4, "{1}", DataType.INT8ARRAY);
-        assertList(fields, 5, null, DataType.FLOAT4ARRAY);
-        assertList(fields, 6, "{1.7E308}", DataType.FLOAT8ARRAY);
-        assertList(fields, 7, null, DataType.TEXTARRAY);
-        assertList(fields, 8, "{\"\\\\xdeadbeef\"}", DataType.BYTEAARRAY);
-        assertList(fields, 9, null, DataType.BPCHARARRAY);
-        assertList(fields, 10, "{hello}", DataType.VARCHARARRAY);
-        assertList(fields, 11, null, DataType.DATEARRAY);
-        assertList(fields, 12, "{" + BigDecimal.valueOf(1234560000000000000L, 18) + "}", DataType.NUMERICARRAY);
-
-        fields = assertRow(groups, 1, 13);
-        assertField(fields, 0, 2, DataType.INTEGER);
-        assertList(fields, 1, null, DataType.BOOLARRAY);
-        assertList(fields, 2, "{-128,96}", DataType.INT2ARRAY);
-        assertList(fields, 3, null, DataType.INT4ARRAY);
-        assertList(fields, 4, "{-9223372036854775808,223372036854775808}", DataType.INT8ARRAY);
-        assertList(fields, 5, null, DataType.FLOAT4ARRAY);
-        assertList(fields, 6, "{1.0,-99.9}", DataType.FLOAT8ARRAY);
-        assertList(fields, 7, null, DataType.TEXTARRAY);
-        assertList(fields, 8, "{\"\\\\xdeadbeef\",\"\\\\xadbeef\"}", DataType.BYTEAARRAY);
-        assertList(fields, 9, null, DataType.BPCHARARRAY);
-        assertList(fields, 10, "{\"this is exactly\",\" fifteen chars.\"}", DataType.VARCHARARRAY);
-        assertList(fields, 11, null, DataType.DATEARRAY);
-        assertList(fields, 12, "{" + BigDecimal.valueOf(1234560000000000000L, 18) + "," + BigDecimal.valueOf(1234560000000000000L, 18) + "}", DataType.NUMERICARRAY);
-    }
-
-    @Test
-    public void testGetFields_Timestamp_List_Nulls() throws IOException {
-        schema = getParquetSchemaForTimestampListTypeGeneratedBySpark();
-        // schema has changed, set metadata again
-        context.setMetadata(schema);
-        context.setTupleDescription(getColumnDescriptorsFromSchema(schema));
-        resolver.setRequestContext(context);
-        resolver.afterPropertiesSet();
-
-        List<Group> groups = readParquetFile("parquet_timestamp_list_type.parquet", 6, schema);
-        assertEquals(6, groups.size());
-
-        // for test cases that test conversions against server's time zone
-        Instant timestamp1 = Instant.parse("2022-10-05T18:30:00Z"); // UTC
-        Instant timestamp2 = Instant.parse("2022-10-06T19:30:00Z"); // UTC
-        Instant timestamp3 = Instant.parse("2022-10-07T20:30:00Z"); // UTC
-        ZonedDateTime localTime1 = timestamp1.atZone(ZoneId.systemDefault());
-        ZonedDateTime localTime2 = timestamp2.atZone(ZoneId.systemDefault());
-        ZonedDateTime localTime3 = timestamp3.atZone(ZoneId.systemDefault());
-        String localTimestampString1 = localTime1.format(DateTimeFormatter.ofPattern("yyyy-MM-dd HH:mm:ss"));
-        String localTimestampString2 = localTime2.format(DateTimeFormatter.ofPattern("yyyy-MM-dd HH:mm:ss"));
-        String localTimestampString3 = localTime3.format(DateTimeFormatter.ofPattern("yyyy-MM-dd HH:mm:ss"));
-
-        List<OneField> fields = assertRow(groups, 0, 2);
-        assertField(fields, 0, 1, DataType.INTEGER);
-        assertList(fields, 1, "{\"" + localTimestampString1 + "\",\"" + localTimestampString2 + "\",\"" + localTimestampString3 + "\"}", DataType.TIMESTAMPARRAY);
-
-        fields = assertRow(groups, 1, 2);
-        assertField(fields, 0, 2, DataType.INTEGER);
-        assertList(fields, 1, "{\"" + localTimestampString1 + "\",\"" + localTimestampString1 + "\",\"" + localTimestampString3 + "\"}", DataType.TIMESTAMPARRAY);
-
-        fields = assertRow(groups, 2, 2);
-        assertField(fields, 0, 3, DataType.INTEGER);
-        assertList(fields, 1, "{NULL,\"" + localTimestampString1 + "\",\"" + localTimestampString1 + "\"}", DataType.TIMESTAMPARRAY);
-
-        fields = assertRow(groups, 3, 2);
-        assertField(fields, 0, 4, DataType.INTEGER);
-        assertList(fields, 1, "{NULL}", DataType.TIMESTAMPARRAY);
-
-        fields = assertRow(groups, 4, 2);
-        assertField(fields, 0, 5, DataType.INTEGER);
-        assertList(fields, 1, "{}", DataType.TIMESTAMPARRAY);
-
-        fields = assertRow(groups, 5, 2);
-        assertField(fields, 0, 6, DataType.INTEGER);
-        assertList(fields, 1, null, DataType.TIMESTAMPARRAY);
-    }
-
-    @Test
-    @SuppressWarnings("deprecation")
-    public void testGetFields_Unsupported_Struct_List() {
-        List<ColumnDescriptor> columnDescriptors = new ArrayList<>();
-        ColumnDescriptor listColumnDescriptor = new ColumnDescriptor("unsupported_list", -1, 1, "", new Integer[]{});
-        columnDescriptors.add(listColumnDescriptor);
-
-        // LIST of customized Struct, with no parquet original type name
-        schema = getParquetSchemaForUnsupportedListType()[0];
-        // schema has changed, set metadata again
-        context.setMetadata(schema);
-        context.setTupleDescription(columnDescriptors);
-        resolver.setRequestContext(context);
-        resolver.afterPropertiesSet();
-
-        List<Type> structFields = new ArrayList<>();
-
-        structFields.add(new PrimitiveType(Type.Repetition.OPTIONAL, PrimitiveTypeName.INT32, "num1"));
-        structFields.add(new PrimitiveType(Type.Repetition.OPTIONAL, PrimitiveTypeName.INT32, "num2"));
-        structFields.add(new PrimitiveType(Type.Repetition.OPTIONAL, PrimitiveTypeName.INT32, "num3"));
-
-        GroupType structGroupType = new GroupType(Type.Repetition.OPTIONAL, "element", structFields);
-        Group structGroup = new SimpleGroup(structGroupType);
-
-        GroupType listRepeatedGroupType = new GroupType(Type.Repetition.REPEATED, "repeated_list", structGroupType);
-        Group listRepeatedGroup = new SimpleGroup(listRepeatedGroupType);
-        listRepeatedGroup.add(0, structGroup);
-
-        GroupType listGroupType = new GroupType(Type.Repetition.OPTIONAL, "unsupported_list", org.apache.parquet.schema.OriginalType.LIST, listRepeatedGroupType);
-        Group listGroup = new SimpleGroup(listGroupType);
-        listGroup.add(0, listRepeatedGroup);
-
-        List<Group> groups = new ArrayList<>();
-        groups.add(listGroup);
-
-        Exception e = assertThrows(UnsupportedTypeException.class,
-                () -> assertRow(groups, 0, 1));
-        assertEquals("Parquet LIST of customized struct is not supported.", e.getMessage());
-    }
-
-    @Test
-    @SuppressWarnings("deprecation")
-    public void testGetFields_Unsupported_Map_List() {
-        List<ColumnDescriptor> columnDescriptors = new ArrayList<>();
-        ColumnDescriptor listColumnDescriptor = new ColumnDescriptor("unsupported_list", -1, 1, "", new Integer[]{});
-        columnDescriptors.add(listColumnDescriptor);
-
-        // LIST of customized Struct, with no parquet original type name
-        schema = getParquetSchemaForUnsupportedListType()[1];
-        // schema has changed, set metadata again
-        context.setMetadata(schema);
-        context.setTupleDescription(columnDescriptors);
-        resolver.setRequestContext(context);
-        resolver.afterPropertiesSet();
-
-        List<Type> mapFields = new ArrayList<>();
-
-        mapFields.add(new PrimitiveType(Type.Repetition.REQUIRED, PrimitiveTypeName.INT32, "key"));
-        mapFields.add(new PrimitiveType(Type.Repetition.OPTIONAL, PrimitiveTypeName.INT32, "value"));
-
-        GroupType mapElementGroupType = new GroupType(Type.Repetition.OPTIONAL, "element", mapFields);
-        Group mapElementGroup = new SimpleGroup(mapElementGroupType);
-
-        GroupType mapRepeatedGroupType = new GroupType(Type.Repetition.REPEATED, "map", mapElementGroupType);
-        Group mapRepeatedGroup = new SimpleGroup(mapRepeatedGroupType);
-        mapRepeatedGroup.add(0, mapElementGroup);
-
-        GroupType mapGroupType = new GroupType(Type.Repetition.OPTIONAL, "my_map", org.apache.parquet.schema.OriginalType.MAP, mapRepeatedGroupType);
-        GroupType listRepeatedGroupType = new GroupType(Type.Repetition.REPEATED, "repeated_list", mapGroupType);
-        Group listRepeatedGroup = new SimpleGroup(listRepeatedGroupType);
-        listRepeatedGroup.add(0, mapRepeatedGroup);
-
-        GroupType listGroupType = new GroupType(Type.Repetition.OPTIONAL, "unsupported_list", org.apache.parquet.schema.OriginalType.LIST, listRepeatedGroupType);
-        Group listGroup = new SimpleGroup(listGroupType);
-        listGroup.add(0, listRepeatedGroup);
-
-        List<Group> groups = new ArrayList<>();
-        groups.add(listGroup);
-
-        Exception e = assertThrows(UnsupportedTypeException.class,
-                () -> assertRow(groups, 0, 1));
-        assertEquals("Parquet LIST of MAP is not supported.", e.getMessage());
-    }
-
-    @Test
-    @SuppressWarnings("deprecation")
-    public void testGetFields_Unsupported_List_List() {
-        List<ColumnDescriptor> columnDescriptors = new ArrayList<>();
-        ColumnDescriptor listColumnDescriptor = new ColumnDescriptor("unsupported_list", -1, 1, "", new Integer[]{});
-        columnDescriptors.add(listColumnDescriptor);
-
-        // LIST of LIST
-        schema = getParquetSchemaForUnsupportedListType()[2];
-        // schema has changed, set metadata again
-        context.setMetadata(schema);
-        columnDescriptors = new ArrayList<>();
-        listColumnDescriptor = new ColumnDescriptor("unsupported_list", -1, 1, "", new Integer[]{});
-        columnDescriptors.add(listColumnDescriptor);
-        context.setTupleDescription(columnDescriptors);
-        resolver.setRequestContext(context);
-        resolver.afterPropertiesSet();
-
-        List<Type> innerListElementFields = new ArrayList<>();
-        innerListElementFields.add(new PrimitiveType(Type.Repetition.REQUIRED, PrimitiveTypeName.INT32, "id"));
-
-        GroupType innerListElementGroupType = new GroupType(Type.Repetition.OPTIONAL, "element", innerListElementFields);
-        Group innerListElementGroup = new SimpleGroup(innerListElementGroupType);
-
-        GroupType innerListRepeatedGroupType = new GroupType(Type.Repetition.REPEATED, "inner_repeated_list", innerListElementGroupType);
-        Group innerRepeatedGroup = new SimpleGroup(innerListRepeatedGroupType);
-        innerRepeatedGroup.add(0, innerListElementGroup);
-
-        GroupType innerListGroupType = new GroupType(Type.Repetition.OPTIONAL, "inner_list", org.apache.parquet.schema.OriginalType.LIST, innerListRepeatedGroupType);
-        Group innerListGroup = new SimpleGroup(innerListGroupType);
-        innerListGroup.add(0, innerRepeatedGroup);
-
-        GroupType listRepeatedGroupType = new GroupType(Type.Repetition.REPEATED, "repeated_list", innerListGroupType);
-        Group listRepeatedGroup = new SimpleGroup(listRepeatedGroupType);
-        listRepeatedGroup.add(0, innerListGroup);
-
-        GroupType listGroupType = new GroupType(Type.Repetition.OPTIONAL, "unsupported_list", org.apache.parquet.schema.OriginalType.LIST, listRepeatedGroupType);
-        Group listGroup = new SimpleGroup(listGroupType);
-        listGroup.add(0, listRepeatedGroup);
-
-        List<Group> groups = new ArrayList<>();
-        groups.add(listGroup);
-
-        Exception e = assertThrows(UnsupportedTypeException.class,
-                () -> assertRow(groups, 0, 1));
-        assertEquals("Parquet LIST of LIST is not supported.", e.getMessage());
-    }
-
-    @Test
-    public void testGetFields_Unsupported_Complex() {
-        schema = getParquetSchemaForUnsupportedComplexType();
-        // schema has changed, set metadata again
-        context.setMetadata(schema);
-        Exception e = assertThrows(UnsupportedTypeException.class,
-                () -> context.setTupleDescription(getColumnDescriptorsFromSchema(schema)));
-        assertEquals("Parquet complex type MAP is not supported, error: java.lang.IllegalArgumentException: No enum constant org.greenplum.pxf.plugins.hdfs.parquet.ParquetTypeConverter.MAP", e.getMessage());
-    }
-
-    @Test
-    @SuppressWarnings("deprecation")
-    public void testGetFields_Invalid_Schema() {
-        // test invalid parquet primitive schema
-        List<Type> fields = new ArrayList<>();
-        fields.add(new PrimitiveType(Type.Repetition.OPTIONAL, null, "", null));
-        schema = new MessageType("invalid_list_schema", fields);
-        context.setMetadata(schema);
-        Exception e = assertThrows(PxfRuntimeException.class,
-                () -> context.setTupleDescription(getColumnDescriptorsFromSchema(schema)));
-        assertEquals("Invalid Parquet primitive schema. Parquet primitive type name is null.", e.getMessage());
-
-        // test invalid parquet list schema
-        fields = new ArrayList<>();
-        fields.add(new GroupType(Type.Repetition.OPTIONAL, "invalid_list", org.apache.parquet.schema.OriginalType.LIST, new ArrayList<>()));
-        schema = new MessageType("invalid_list_schema", fields);
-        context.setMetadata(schema);
-        e = assertThrows(PxfRuntimeException.class,
-                () -> context.setTupleDescription(getColumnDescriptorsFromSchema(schema)));
-        assertEquals(String.format("Invalid Parquet List schema: %s.", schema.getFields().get(0).toString()), e.getMessage());
-
-        fields = new ArrayList<>();
-        GroupType repeatedGroupType = new GroupType(Type.Repetition.REPEATED, "list", new ArrayList<>());
-        fields.add(new GroupType(Type.Repetition.OPTIONAL, "invalid_list", org.apache.parquet.schema.OriginalType.LIST, repeatedGroupType));
-        schema = new MessageType("invalid_list_schema", fields);
-        context.setMetadata(schema);
-        e = assertThrows(PxfRuntimeException.class,
-                () -> context.setTupleDescription(getColumnDescriptorsFromSchema(schema)));
-        assertEquals(String.format("Invalid Parquet List schema: %s.", schema.getFields().get(0).toString()), e.getMessage());
-    }
-
     private List<OneField> assertRow(List<Group> groups, int desiredRow, int numFields) {
         OneRow row = new OneRow(groups.get(desiredRow)); // get row
         List<OneField> fields = resolver.getFields(row);
@@ -874,29 +528,12 @@ public class ParquetResolverTest {
         return fields;
     }
 
-    private void assertField(List<OneField> fields, int index, Object expectedValue, DataType type) {
+    private void assertField(List<OneField> fields, int index, Object value, DataType type) {
         assertEquals(type.getOID(), fields.get(index).type);
-
-        if (expectedValue == null) {
-            assertNull(fields.get(index).val);
-        } else if (type == DataType.BYTEA) {
-            assertArrayEquals((byte[]) expectedValue, (byte[]) fields.get(index).val);
+        if (type == DataType.BYTEA) {
+            assertArrayEquals((byte[]) value, (byte[]) fields.get(index).val);
         } else {
-            assertEquals(expectedValue, fields.get(index).val);
-        }
-    }
-
-    private void assertList(List<OneField> fields, int index, String expectedValue, DataType type) {
-        if (type.getOID() == DataType.BPCHARARRAY.getOID() || type.getOID() == DataType.VARCHARARRAY.getOID()) {
-            assertEquals(DataType.TEXTARRAY.getOID(), fields.get(index).type);
-        } else {
-            assertEquals(type.getOID(), fields.get(index).type);
-        }
-
-        if (expectedValue == null) {
-            assertNull(fields.get(index).val);
-        } else {
-            assertEquals(expectedValue, String.valueOf(fields.get(index).val));
+            assertEquals(value, fields.get(index).val);
         }
     }
 
@@ -929,114 +566,6 @@ public class ParquetResolverTest {
     }
 
     @SuppressWarnings("deprecation")
-    private MessageType getParquetSchemaForListTypesGeneratedByHive() {
-        List<Type> fields = new ArrayList<>();
-
-        Type.Repetition groupRepetition = Type.Repetition.OPTIONAL;
-        Type.Repetition elementRepetition = Type.Repetition.OPTIONAL;
-        fields.add(new PrimitiveType(elementRepetition, PrimitiveTypeName.INT32, "id", null));
-        fields.add(generateListSchema(groupRepetition, "bag", elementRepetition, "array_element", PrimitiveTypeName.BOOLEAN, 0, "bool_arr", null));
-        org.apache.parquet.schema.OriginalType tinyType = org.apache.parquet.schema.OriginalType.INT_8;
-        fields.add(generateListSchema(groupRepetition, "bag", elementRepetition, "array_element", PrimitiveTypeName.INT32, 0, "smallint_arr", tinyType));
-        fields.add(generateListSchema(groupRepetition, "bag", elementRepetition, "array_element", PrimitiveTypeName.INT32, 0, "int_arr", null));
-        fields.add(generateListSchema(groupRepetition, "bag", elementRepetition, "array_element", PrimitiveTypeName.INT64, 0, "bigint_arr", null));
-        fields.add(generateListSchema(groupRepetition, "bag", elementRepetition, "array_element", PrimitiveTypeName.FLOAT, 0, "real_arr", null));
-        fields.add(generateListSchema(groupRepetition, "bag", elementRepetition, "array_element", PrimitiveTypeName.DOUBLE, 0, "double_arr", null));
-        fields.add(generateListSchema(groupRepetition, "bag", elementRepetition, "array_element", PrimitiveTypeName.BINARY, 0, "text_arr", org.apache.parquet.schema.OriginalType.UTF8));
-        fields.add(generateListSchema(groupRepetition, "bag", elementRepetition, "array_element", PrimitiveTypeName.BINARY, 0, "bytea_arr", null));
-        fields.add(generateListSchema(groupRepetition, "bag", elementRepetition, "array_element", PrimitiveTypeName.BINARY, 0, "char_arr", org.apache.parquet.schema.OriginalType.UTF8));
-        fields.add(generateListSchema(groupRepetition, "bag", elementRepetition, "array_element", PrimitiveTypeName.BINARY, 0, "varchar_arr", org.apache.parquet.schema.OriginalType.UTF8));
-        fields.add(generateListSchema(groupRepetition, "bag", elementRepetition, "array_element", PrimitiveTypeName.INT32, 0, "date_arr", org.apache.parquet.schema.OriginalType.DATE));
-        fields.add(generateListSchema(groupRepetition, "bag", elementRepetition, "array_element", PrimitiveTypeName.FIXED_LEN_BYTE_ARRAY, 16, "numeric_arr", org.apache.parquet.schema.OriginalType.DECIMAL));
-
-        return new MessageType("hive_schema", fields);
-    }
-
-    @SuppressWarnings("deprecation")
-    private MessageType getParquetSchemaForTimestampListTypeGeneratedBySpark() {
-        List<Type> fields = new ArrayList<>();
-        fields.add(new PrimitiveType(Type.Repetition.OPTIONAL, PrimitiveTypeName.INT32, "id", null));
-        fields.add(generateListSchema(Type.Repetition.OPTIONAL, "list", Type.Repetition.OPTIONAL, "element", PrimitiveTypeName.INT96, 0, "tm_arr", null));
-        return new MessageType("spark_schema", fields);
-    }
-
-    @SuppressWarnings("deprecation")
-    private GroupType generateListSchema(Type.Repetition groupRepetition, String repeatedGroupName, Type.Repetition elementRepetition, String elementName, PrimitiveTypeName primitiveTypeName, int length, String listName, org.apache.parquet.schema.OriginalType originalType) {
-        PrimitiveType elementType;
-        if (originalType == org.apache.parquet.schema.OriginalType.DECIMAL) {
-            elementType = new PrimitiveType(elementRepetition, primitiveTypeName, length, elementName, originalType, new org.apache.parquet.schema.DecimalMetadata(38, 18), null);
-        } else {
-            elementType = new PrimitiveType(elementRepetition, primitiveTypeName, elementName, originalType);
-        }
-        GroupType repeatedGroupType = new GroupType(Type.Repetition.REPEATED, repeatedGroupName, elementType);
-        return new GroupType(groupRepetition, listName, org.apache.parquet.schema.OriginalType.LIST, repeatedGroupType);
-    }
-
-    @SuppressWarnings("deprecation")
-    private MessageType[] getParquetSchemaForUnsupportedListType() {
-        MessageType[] messageTypes = new MessageType[3];
-
-        // List of customized Struct
-        List<Type> structFields = new ArrayList<>();
-
-        structFields.add(new PrimitiveType(Type.Repetition.OPTIONAL, PrimitiveTypeName.INT32, "num1"));
-        structFields.add(new PrimitiveType(Type.Repetition.OPTIONAL, PrimitiveTypeName.INT32, "num2"));
-        structFields.add(new PrimitiveType(Type.Repetition.OPTIONAL, PrimitiveTypeName.INT32, "num3"));
-
-        GroupType structGroupType = new GroupType(Type.Repetition.OPTIONAL, "element", structFields);
-        GroupType listRepeatedGroupType = new GroupType(Type.Repetition.REPEATED, "repeated_list", structGroupType);
-        GroupType listGroupType = new GroupType(Type.Repetition.OPTIONAL, "unsupported_list", org.apache.parquet.schema.OriginalType.LIST, listRepeatedGroupType);
-        List<Type> fields = new ArrayList<>();
-        fields.add(listGroupType);
-        messageTypes[0] = new MessageType("spark_schema", fields);
-
-        // List of Map
-        List<Type> mapFields = new ArrayList<>();
-
-        mapFields.add(new PrimitiveType(Type.Repetition.REQUIRED, PrimitiveTypeName.INT32, "key"));
-        mapFields.add(new PrimitiveType(Type.Repetition.OPTIONAL, PrimitiveTypeName.INT32, "value"));
-
-        GroupType mapRepeatedGroupType = new GroupType(Type.Repetition.REPEATED, "map", mapFields);
-        GroupType mapGroupType = new GroupType(Type.Repetition.OPTIONAL, "my_map", org.apache.parquet.schema.OriginalType.MAP, mapRepeatedGroupType);
-        listRepeatedGroupType = new GroupType(Type.Repetition.REPEATED, "repeated_list", mapGroupType);
-        listGroupType = new GroupType(Type.Repetition.OPTIONAL, "unsupported_list", org.apache.parquet.schema.OriginalType.LIST, listRepeatedGroupType);
-        fields = new ArrayList<>();
-
-        fields.add(listGroupType);
-        messageTypes[1] = new MessageType("spark_schema", fields);
-
-        // List of List
-        List<Type> innerListElementFields = new ArrayList<>();
-
-        innerListElementFields.add(new PrimitiveType(Type.Repetition.REQUIRED, PrimitiveTypeName.INT32, "id"));
-
-        GroupType innerListRepeatedGroupType = new GroupType(Type.Repetition.REPEATED, "inner_repeated_list", innerListElementFields);
-        GroupType innerListGroupType = new GroupType(Type.Repetition.OPTIONAL, "inner_list", org.apache.parquet.schema.OriginalType.LIST, innerListRepeatedGroupType);
-        listRepeatedGroupType = new GroupType(Type.Repetition.REPEATED, "repeated_list", innerListGroupType);
-        listGroupType = new GroupType(Type.Repetition.OPTIONAL, "unsupported_list", org.apache.parquet.schema.OriginalType.LIST, listRepeatedGroupType);
-        fields = new ArrayList<>();
-
-        fields.add(listGroupType);
-        messageTypes[2] = new MessageType("spark_schema", fields);
-        return messageTypes;
-    }
-
-    @SuppressWarnings("deprecation")
-    private MessageType getParquetSchemaForUnsupportedComplexType() {
-        List<Type> mapFields = new ArrayList<>();
-
-        mapFields.add(new PrimitiveType(Type.Repetition.REQUIRED, PrimitiveTypeName.INT32, "key"));
-        mapFields.add(new PrimitiveType(Type.Repetition.OPTIONAL, PrimitiveTypeName.INT32, "value"));
-
-        GroupType repeatedGroupType = new GroupType(Type.Repetition.REPEATED, "map", mapFields);
-        GroupType groupType = new GroupType(Type.Repetition.OPTIONAL, "unsupported_complex", org.apache.parquet.schema.OriginalType.MAP, repeatedGroupType);
-        List<Type> fields = new ArrayList<>();
-
-        fields.add(groupType);
-        return new MessageType("spark_schema", fields);
-    }
-
-    @SuppressWarnings("deprecation")
     private List<Group> readParquetFile(String file, long expectedSize, MessageType schema) throws IOException {
         List<Group> result = new ArrayList<>();
         String parquetFile = Objects.requireNonNull(getClass().getClassLoader().getResource("parquet/" + file)).getPath();
@@ -1061,8 +590,7 @@ public class ParquetResolverTest {
         return schema.getFields()
                 .stream()
                 .map(f -> {
-                    Type type = f.isPrimitive() ? f.asPrimitiveType() : f.asGroupType();
-                    ParquetTypeConverter converter = ParquetTypeConverter.from(type);
+                    ParquetTypeConverter converter = ParquetTypeConverter.from(f.asPrimitiveType());
                     return new ColumnDescriptor(f.getName(), converter.getDataType(f).getOID(), 1, "", new Integer[]{});
                 })
                 .collect(Collectors.toList());
