--- server/pxf-hdfs/src/test/resources/parquet/README.md
@@ -1,4 +1,0 @@
-# Generate Parquet files for testing
-
-The parquet files are pre-generated, but if you want to generate these files again,
-please refer to this [README](../../../../../../automation/src/test/resources/data/parquet/README.md)
